requirements.txt
diffusers
transformers
torch
accelerate
safetensors
Pillow
requests
rembg
Flask
python-dotenv